Make sure UI state matches DB state.
If a quantitative question is configured to display options in random order, the browser will sometimes keep certain options checked after refreshing the page without clearing the cache. This happens for options that are unchecked when the learner submits their answers, and that switch places with checked options (due to randomization) after a regular page refresh. So we need to make sure that the rendered state of individual options matches the state of the DB (which is represented in the raw HTML that the LPD view generates).